me quede en agregar stock

+++ src/main/java/com/mycompany/supermecardo/entidades/Producto.java
@@ -10,17 +10,19 @@ public class Producto implements Serializable {
     @Id
     private String codigoId;
     private String nombre;
-    private Integer stock;
+    private Double stock;
     private Integer costo;
     private Integer precio;
     private String categoria;
     private String fromVenta;
-    private Integer unidadesVendidas;
+    private Double unidadesVendidas;
     
     public Producto() {
     }
 
-    public Producto(String codigoId, String nombre, Integer stock, Integer costo, Integer precio, String categoria, String fromVenta, Integer unidadesVendidas) {
+    public Producto(String codigoId, String nombre, Double stock,
+            Integer costo, Integer precio, String categoria, 
+            String fromVenta, Double unidadesVendidas) {
         this.codigoId = codigoId;
         this.nombre = nombre;
         this.stock = stock;
@@ -47,13 +49,15 @@ public class Producto implements Serializable {
         this.nombre = nombre;
     }
 
-    public Integer getStock() {
+    public Double getStock() {
         return stock;
     }
 
-    public void setStock(Integer stock) {
+    public void setStock(Double stock) {
         this.stock = stock;
     }
+
+
 
     public Integer getCosto() {
         return costo;
@@ -87,11 +91,11 @@ public class Producto implements Serializable {
         this.fromVenta = fromVenta;
     }
 
-    public Integer getUnidadesVendidas() {
+    public Double getUnidadesVendidas() {
         return unidadesVendidas;
     }
 
-    public void setUnidadesVendidas(Integer unidadesVendidas) {
+    public void setUnidadesVendidas(Double unidadesVendidas) {
         this.unidadesVendidas = unidadesVendidas;
     }
